arcID to spatID

+++ src/Simulation/Stepper.hpp
@@ -78,8 +78,8 @@ namespace GlobalFlow {
                     _delta_t_n = __delta;
 
                     LOG(debug) << "Stepsize delta " << __delta;
-                    LOG(debug) << "Stepsize: " << _time *  __delta;
-                    _stepper->get(0)->updateStepSize(_time * __delta);
+                    LOG(debug) << "Stepsize: " << double(_time) *  __delta;
+                    _stepper->get(0)->updateStepSize(double(_time) * __delta);
                     _pos = _pos + __delta;
                     LOG(debug) << "Current position " << _pos;
 
@@ -93,8 +93,8 @@ namespace GlobalFlow {
                 double __delta{0};
                 __delta = _totalSteps * ((_p - 1) / (std::pow(_p, _totalSteps) - 1));
                 _delta_t_n = __delta;
-                _stepper->get(0)->updateStepSize(_time * __delta);
-		            LOG(debug) << "Stepsize: " << _time * __delta;
+                _stepper->get(0)->updateStepSize(double(_time) * __delta);
+		            LOG(debug) << "Stepsize: " << double(_time) * __delta;
                 _pos = _pos + __delta;
             }
